setup maze and silver maze helper
